Reduce the amount of GenServers to 1. State is contained in 1 Server.
I think this is an advancement to previous design because GenServers
are being used more intentionally. Socket for example didn't have to
be a server.

I feel there's a process message bottleneck. The main server is processing
messages related to status reads, announcement reads and user messages.
In case any of the reading fails, the whole server goes away and all the
state has been lost. You are forced to respawn the whole system.

Calling .query() the first time takes a while to run, I don't know why.