Uses 2 main processes to gather information from the network.
This design feels more idiomatic and I'm more satisfied with it.
A clunky thing that still annoys me is the way the keep alive packets
are being created. The need to be passing an Iface feels off too.

Before the process gets linked, a new dictionary if created with
vcdj properties that will stick around just so the keep alive
packet can be created at the moment of sending. This recreates the
packet every, single, time. I don't like that since there's really
no need for the packet to be regenerated.

I'm trying to find a way in which this annoyance gets fixed and I'm
thinking there could probably be another process that handles the
specifics of network interactions. Not sure. This idea takes
inspiration on the fact that separating process by the socket they
use was probably a good move.

Another challenge I'll have to resolve is that I'm wrapping UDP sockets
with `Socket`. Later on I'll need TCP sockets to query for the current
playing track so the abstraction will need to change.