Packet now parses data after a data structure with the packet information. This is alleviating so much work from doing packet parsin by hand.
Next step is to implement the rekordbox server.
In order to start it, I need a device ip so a tcp connection can be made.
It's possible to decide from a list, and have the cdj map a bool indicating
if it's a virtual or real device.

Grab any real device, get the ip, and make the connection.

Once the result comes back, make sure to keep the necessary information
in the process state so later call can be made such as getting the track
metadata, artwork and more.