Makes VCDJ the main process that controls the others.
I chose to split processes by the information gathered
on each port. This became problematic when I got to the
point of asking for track metadata since it required
to open a TCP connection. So far the idea of port for
the system was a UDP port. This approach felt better, but
isn't working.

I opted to now create a main process (VCDJ) which acts
as if it was a physical device with processes under it.
This isn't properly supervised yet.

Data strucutres.

Data is received from 2 different ports and is held at
2 different processes. It feels redundant to have status
and presence data on different places.

I think a better over arching data structure can be used
for this program.